Se separa declaracion e implementacion de clase ReproductorSonido en archivos diferentes .h y .cpp
